Backend/Analytics_Service/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic 
psycopg2-binary